requirements.txt
Flask==2.2.1
Flask-MySQL==1.5.2
mysql-connector-python==8.0.30
